test: group provider-specific isvctl tests by provider
