+++ isvctl/tests/providers/aws/test_aws_control_plane_scripts.py
@@ -26,7 +26,7 @@ from typing import Any
 
 import pytest
 
-ISVCTL_ROOT = Path(__file__).resolve().parents[1]
+ISVCTL_ROOT = Path(__file__).resolve().parents[3]
 AWS_CONTROL_PLANE_SCRIPTS = ISVCTL_ROOT / "configs" / "providers" / "aws" / "scripts" / "control-plane"
