+++ isvctl/tests/providers/aws/test_aws_eks_shared_vpc_cluster_scripts.py
@@ -13,7 +13,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Contract tests for AWS EKS node-pool helper scripts."""
+"""Contract tests for AWS EKS shared-VPC cluster helper scripts."""
+
+from __future__ import annotations
 
 import os
 import subprocess
@@ -21,14 +23,14 @@ from pathlib import Path
 
 import pytest
 
-AWS_EKS_DIR = Path(__file__).resolve().parents[1] / "configs" / "providers" / "aws" / "scripts" / "eks"
+AWS_EKS_DIR = Path(__file__).resolve().parents[3] / "configs" / "providers" / "aws" / "scripts" / "eks"
 
 
-@pytest.mark.parametrize("script_name", ["create_node_pool.sh", "destroy_node_pool.sh"])
-@pytest.mark.parametrize("state_file", ["../terraform.tfstate", "/tmp/node-pool.tfstate"])
-def test_node_pool_scripts_reject_state_file_paths(script_name: str, state_file: str) -> None:
-    """Node-pool state overrides must stay local to terraform-node-pool/."""
-    env = {**os.environ, "NODE_POOL_STATE_FILE": state_file}
+@pytest.mark.parametrize("script_name", ["create_shared_vpc_cluster.sh", "destroy_shared_vpc_cluster.sh"])
+@pytest.mark.parametrize("state_file", ["../terraform.tfstate", "/tmp/shared-vpc-cluster.tfstate"])
+def test_shared_vpc_cluster_scripts_reject_state_file_paths(script_name: str, state_file: str) -> None:
+    """Shared-cluster state overrides must stay local to terraform-shared-vpc-cluster/."""
+    env = {**os.environ, "SHARED_VPC_CLUSTER_STATE_FILE": state_file}
 
     completed = subprocess.run(
         ["bash", str(AWS_EKS_DIR / script_name)],
@@ -39,4 +41,4 @@ def test_node_pool_scripts_reject_state_file_paths(script_name: str, state_file:
     )
 
     assert completed.returncode == 1
-    assert "NODE_POOL_STATE_FILE must be a local .tfstate filename" in completed.stderr
+    assert "SHARED_VPC_CLUSTER_STATE_FILE must be a local .tfstate filename" in completed.stderr
