+++ isvctl/tests/providers/aws/test_aws_ssh_utils.py
@@ -25,7 +25,7 @@ from typing import Any
 
 import pytest
 
-ISVCTL_ROOT = Path(__file__).resolve().parents[1]
+ISVCTL_ROOT = Path(__file__).resolve().parents[3]
 AWS_COMMON_SCRIPTS = ISVCTL_ROOT / "configs" / "providers" / "aws" / "scripts" / "common"
