+++ isvctl/tests/providers/aws/test_aws_verified_reuse.py
@@ -31,7 +31,7 @@ from unittest.mock import MagicMock
 import pytest
 from botocore.exceptions import ClientError
 
-_COMMON_DIR = Path(__file__).resolve().parents[1] / "configs" / "providers" / "aws" / "scripts"
+_COMMON_DIR = Path(__file__).resolve().parents[3] / "configs" / "providers" / "aws" / "scripts"
 if str(_COMMON_DIR) not in sys.path:
     sys.path.insert(0, str(_COMMON_DIR))
